move ArtifactData to containers
