+++ qiskit_experiments/framework/base_analysis.py
@@ -23,7 +23,7 @@ from dateutil import tz
 
 from qiskit_experiments.database_service.device_component import Qubit
 from qiskit_experiments.framework import Options
-from qiskit_experiments.framework.artifact_data import ArtifactData
+from qiskit_experiments.framework.containers.artifact_data import ArtifactData
 from qiskit_experiments.framework.store_init_args import StoreInitArgs
 from qiskit_experiments.framework.experiment_data import ExperimentData, FigureData
 from qiskit_experiments.framework.configs import AnalysisConfig
