+++ qiskit_experiments/framework/containers/__init__.py
@@ -10,3 +10,5 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
+from .artifact_data import ArtifactData
+from .figure_data import FigureData
